+++ docs/assets/chevron-up-CVRLtbLB.js
@@ -1,1 +1,0 @@
-import{gt as e}from"./index-CmXInfwf.js";var t=e(`chevron-up`,[[`path`,{d:`m18 15-6-6-6 6`,key:`153udz`}]]);export{t};